Reorganized folder structure, by moving route Trade into views folder and created a components folder to hold the Trade views components. Also, created a component for the Header and placed layout structure into a folder.
