+++ package.json
@@ -12,6 +12,7 @@
     "core-js": "^3.6.4",
     "vue": "^2.6.11",
     "vue-router": "^3.1.6",
+    "vue-to-react": "^1.0.0",
     "vuex": "^3.1.3"
   },
   "devDependencies": {
