+++ src/components/ftbot/Balance.vue
