+++ src/components/ftbot/BotAlerts.vue
