+++ src/components/ftbot/BotControls.vue
