+++ src/components/ftbot/BotStatus.vue
