+++ src/components/ftbot/ForceBuyForm.vue
