+++ src/components/ftbot/Performance.vue
