+++ src/components/ftbot/TradeList.vue
