+++ src/components/ftbot/TradeView.vue
